backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
gunicorn>=21.2.0
agno>=0.1.0
openai>=1.12.0
stripe>=8.0.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.27
alembic>=1.13.1
psycopg2-binary>=2.9.9
asyncpg>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.1.0
python-multipart
twilio>=9.0.0
requests>=2.31.0
aiohttp>=3.9.3
livekit-agents
livekit-api>=0.1.0
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-elevenlabs
livekit-plugins-silero
livekit-plugins-tavus
deepgram-sdk>=3.1.0
cryptography>=42.0.0
elevenlabs>=0.2.27
pinecone>=3.1.0
celery>=5.3.6
redis>=5.0.1
pyjwt>=2.8.0
google-cloud-storage
boto3
google-auth-oauthlib==1.2.3
google-api-python-client==2.187.0
pypdf
httpx>=0.27.0
# Force rebuild 2
beautifulsoup4>=4.12.0
nanoid>=2.0.0

asyncari>=0.1.0
PyPDF2
livekit-plugins-xai
resend>=0.1.0
tavily-python>=0.5.0


caldav>=1.3.9
